# src/swagger.yaml
openapi: 3.0.0
info:
  title: Student Attendance API
  version: 1.0.0
  description: API for managing classes and student attendance with face recognition
paths:
...
  /api/auth/status:
    get:
      summary: Get user authentication status
      tags:
        - Auth
      responses:
        '200':
          description: User is authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  isAuthenticated:
                    type: boolean
                    example: true
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 12345
                      role:
                        type: string
                        example: student
        '401':
          description: Unauthorized
...
  /api/class/create:
    post:
      summary: Create a new class
      tags:
        - Class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                className:
                  type: string
                  example: Math 101
                description:
                  type: string
                  example: Basic math class
              required:
                - className
      responses:
        '201':
          description: Class created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          description: Unauthorized
        '403':
          description: Only teachers can create classes
        '400':
          description: Invalid input
...